Rename MomentsApiProvider types
