--- packages/providers/src/ports/MomentsApiProvider/Types/index.ts
@@ -1,2 +1,0 @@
-export * from './CreateMomentInput';
-export * from './CreateMomentResponse';
