+++ packages/providers/src/ports/MomentsApiProvider/index.ts
@@ -1,2 +1,2 @@
-export * from './Types';
+export { CreateMomentInput, CreateMomentResponse } from './types';
 export { MomentsApiProvider } from './MomentsApiProvider';
