+++ packages/providers/src/ports/MomentsApiProvider/types/index.ts
@@ -1,0 +1,2 @@
+export { CreateMomentInput } from './CreateMomentInput';
+export { CreateMomentResponse } from './CreateMomentResponse';
